+++ src/survive_sensor_activations.c
@@ -189,7 +189,7 @@ FLT SurviveSensorActivations_difference(const SurviveSensorActivations *rhs, con
 			}
 		}
 	}
-	return rtn / (double)cnt;
+	return rtn / (FLT)cnt;
 }
 
 SURVIVE_EXPORT uint32_t SurviveSensorActivations_default_tolerance =
